Split all the tests up
