+++ src/examples/regex/Pattern.php
@@ -2,7 +2,7 @@
 // Each Pattern is an alternation between several "Concs"
 // This is the top-level Pattern object returned by the lexer.
 
-namespace Ferno\Loco;
+namespace Ferno\Loco\examples\regex;
 
 class Pattern
 {
